q1/pyproject.toml
[project]
name = "q2"
version = "0.1.0"
description = "Document Analyzer MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.9.2",
    "textblob>=0.17.1",
    "nltk>=3.8.1",
    "py-readability-metrics>=1.4.5",
    "pydantic>=2.5.0"
]
